requirements.txt
discord.py
aiofiles
beautifulsoup4
requests
